+++ tests/e2e/test_runtime.py
@@ -1,6 +1,7 @@
 """
 Module with tests for general CLI runtime
 """
+
 import logging
 import signal
 import subprocess