+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Changed
 ### Added
 - Log stacktrace with source filename, function and line number
+- Log relevant version information at startup of worker
 ### Fixed